Client: Split client id and discovery functions
